clean up and minor refactoring
