+++ src/API/LeadershiProfileAPI.Tests/Infrastructure/MockLogger.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using LeadershipProfileAPI.Controllers;
+using Microsoft.Extensions.Logging;
+
+namespace LeadershipProfileAPI.Tests.Infrastructure
+{
+    public class MockLogger : ILogger<ProfileController>
+    {
+        public IDisposable BeginScope<TState>(TState state)
+        {
+            throw new NotImplementedException();
+        }
+
+        public bool IsEnabled(LogLevel logLevel)
+        {
+            throw new NotImplementedException();
+        }
+
+        public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception exception, Func<TState, Exception, string> formatter)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
