+++ src/API/LeadershipProfileAPI/Constants.cs
@@ -1,0 +1,7 @@
+﻿namespace LeadershipProfileAPI
+{
+    public class Constants
+    {
+        public const string VersionUriFragment = "v5.0.0/api/data/v3";
+    }
+}
