+++ src/API/LeadershipProfileAPI/LeadershipProfileAPI.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30709.132
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeadershipProfileAPI", "LeadershipProfileAPI.csproj", "{F69973E8-264D-43C1-8A83-32FDA0BD47EB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LeadershipProfileAPI", "LeadershipProfileAPI.csproj", "{F69973E8-264D-43C1-8A83-32FDA0BD47EB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeadershiProfileAPI.Tests", "..\LeadershiProfileAPI.Tests\LeadershiProfileAPI.Tests.csproj", "{A183970E-ACF7-46F1-BE9A-1FABBB2EF6A9}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LeadershipProfileAPI.Tests", "..\LeadershiProfileAPI.Tests\LeadershipProfileAPI.Tests.csproj", "{A183970E-ACF7-46F1-BE9A-1FABBB2EF6A9}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
